Readme-gen/All-repo/Dhvani/python-server/requirements.txt
flask
flask-cors
torch
torchvision
numpy
Pillow
huggingface_hub
timm
opencv-python
matplotlib
grad-cam
